Refactor project calendars to shared modules
