+++ frontend/src/dashboard/project/components/Shared/calendar/types.ts
@@ -1,9 +1,7 @@
-import type { TimelineEvent as ApiTimelineEvent } from "@/shared/utils/api";
-
 export type TimelineEvent = {
   id: string;
   eventId?: string;
-  date: string; // YYYY-MM-DD
+  date: string;
   description?: string;
   hours?: number | string;
   budgetItemId?: string | null;
@@ -30,41 +28,3 @@ export type Project = {
   clientPhone?: string;
   clientEmail?: string;
 };
-
-export type CategoryOption =
-  | "AUDIO-VISUAL"
-  | "CLIENT-SERVICES-VIP"
-  | "CONTINGENCY-MISC"
-  | "DECOR"
-  | "DESIGN"
-  | "FABRICATION"
-  | "FOOD-BEVERAGE"
-  | "GRAPHICS"
-  | "INSTALLATION-MATERIALS"
-  | "LABOR"
-  | "LIGHTING"
-  | "MERCH-SWAG"
-  | "PARKING-FUEL-TOLLS"
-  | "PERMITS-INSURANCE"
-  | "PRODUCTION-MGMT"
-  | "RENTALS"
-  | "STORAGE"
-  | "TECH-INTERACTIVES"
-  | "TRAVEL"
-  | "TRUCKING"
-  | "VENUE-LOCATION-FEES"
-  | "WAREHOUSE";
-
-export type UnitOption =
-  | "Each"
-  | "Hrs"
-  | "Days"
-  | "EA"
-  | "PCS"
-  | "Box"
-  | "LF"
-  | "SQFT"
-  | "KG";
-
-// Re-export API type for convenience since the backend response matches TimelineEvent.
-export type { ApiTimelineEvent };
